// src/types/enums.ts
export enum SupportedIconsEnum {
  PERSON_OUTLINE = 'PERSON_OUTLINE',
  EVENT_NOTE = 'EVENT_NOTE',
  FACT_CHECK_OUTLINED = 'FACT_CHECK_OUTLINED',
  INFO_OUTLINED = 'INFO_OUTLINED',
  LOCAL_OFFER_OUTLINED = 'LOCAL_OFFER_OUTLINED',
  BUILD_OUTLINED = 'BUILD_OUTLINED',
  ANCHOR_OUTLINED = 'ANCHOR_OUTLINED',
  HELP_OUTLINE_OUTLINED = 'HELP_OUTLINE_OUTLINED',
  HEADSET_MIC_OUTLINED = 'HEADSET_MIC_OUTLINED',
  WAVES_OUTLINED = 'WAVES_OUTLINED',
  DIRECTIONS_BIKE = 'DIRECTIONS_BIKE',
  BOAT = 'BOAT',
  KAYAK = 'KAYAK',
  CABIN = 'CABIN',
  CAMPING = 'CAMPING',
  CAT = 'CAT',
  MAXIMIZE = 'MAXIMIZE',
  PEOPLE = 'PEOPLE',
  VELOCITY = 'VELOCITY',
  GOOGLE = 'GOOGLE',
  FACEBOOK = 'FACEBOOK',
  INSTAGRAM = 'INSTAGRAM',
  APPLE = 'APPLE',
  OUTLINED_CALL = 'OUTLINED_CALL',
  OUTLINED_CHAT = 'OUTLINED_CHAT',
  OUTLINED_FACEBOOK = 'OUTLINED_FACEBOOK',
  OUTLINED_INSTAGRAM = 'OUTLINED_INSTAGRAM',
  ACCOUNT_CIRCLE_OUTLINED = 'ACCOUNT_CIRCLE_OUTLINED',
  EMAIL_OUTLINED = 'EMAIL_OUTLINED',
  PHONE_ANDROID_OUTLINED = 'PHONE_ANDROID_OUTLINED',
  CALENDAR_TODAY_OUTLINED = 'CALENDAR_TODAY_OUTLINED',
  DRAG_INDICATOR = 'DRAG_INDICATOR',
  LOG_OUT = 'LOG_OUT',
  LOCATION_OUTLINED = 'LOCATION_OUTLINED',
  CALENDAR_TODAY = 'CALENDAR_TODAY',
  TUNE_OUTLINED = 'TUNE_OUTLINED',
}

export enum CardComponentOption {
  DESCRIPTION = 'DESCRIPTION',
  AMENITIES = 'AMENITIES',
}

export enum CategoryType {
  ALL = 'ALL',
  BOATS = 'BOATS',
  KAYAKS = 'KAYAKS',
  CABINS = 'CABINS',
  CAMPING = 'CAMPING',
  BIKES = 'BIKES',
}

export enum BookingInformationType {
  DURATION = 'DURATION',
  PREDEFINED = 'PREDEFINED',
  CUSTOM_RANGE = 'CUSTOM_RANGE',
}

export enum TimeRangeType {
  DURATION = 'DURATION',
  SLIDER_RANGE = 'SLIDER_RANGE',
}

export enum PageType {
  INDEX = 'INDEX',
  ITEM_DETAIL = 'ITEM_DETAIL',
  DOCK_QUEUE = 'DOCK_QUEUE',
  CHECKOUT = 'CHECKOUT',
  MY_RESERVATIONS = 'MY_RESERVATIONS',
}

export enum TagColor {
  RED = '#FF6060',
  LIGHTBLUE = '#1797DA',
  GREEN = '#2DD28D;',
  WHITE = '#FFFFFF',
}

export enum PredefinedCardTags {
  MANDATORY = 'Mandatory',
  RECOMMENDED = 'Recommended',
  POPULAR = 'Popular',
}

export enum CheckoutStepsEnum {
  REGISTER = 'REGISTER',
  TAC = 'TAC',
  CHECKOUT = 'CHECKOUT',
}

export enum CheckoutMethod {
  ADD_TO_CART = 'ADD_TO_CART',
  BOOK_NOW = 'BOOK_NOW',
}

export enum PoliciesReplaceOptionEnum {
  INITIALS = '%your_initials_input%',
  SIGNATURE = '%your_signature_input%',
}

export enum InputType {
  STRING = 'STRING',
  NUMBER = 'NUMBER',
  PHONE = 'PHONE',
  DATE = 'DATE',
  EMAIL = 'EMAIL',
  SELECT = 'SELECT',
  CHECKBOX = 'CHECKBOX',
  RADIO = 'RADIO',
  SINGLECHECK = 'SINGLECHECK',
  TIME = 'TIME',
}

export enum TimePeriod {
  AM = 'AM',
  PM = 'PM',
}

export enum LoginContentInnerStep {
  LOGIN_FORM = 'LOGIN_FORM',
  CUSTOM_INFORMATION = 'CUSTOM_INFORMATION',
}

export enum BrowserType {
  APPLE = 'APPLE',
  GOOGLE = 'GOOGLE',
}